[mtouch/mmp] Unify Xcode lookup & verification code in mtouch and mmp.
Share all the code to find and verify Xcode between mtouch and mmp.

Also print out the actual product version when logging which Xcode was used
(to make it easier to detect if a beta version is in use).
