+++ tools/common/Driver.cs
@@ -16,6 +16,7 @@ using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading.Tasks;
 
+using Xamarin.MacDev;
 using Xamarin.Utils;
 using ObjCRuntime;
 
@@ -23,6 +24,8 @@ namespace Xamarin.Bundler {
 	public partial class Driver {
 		static void AddSharedOptions (Application app, Mono.Options.OptionSet options)
 		{
+			options.Add ("sdkroot=", "Specify the location of Apple SDKs, default to 'xcode-select' value.", v => sdk_root = v);
+			options.Add ("no-xcode-version-check", "Ignores the Xcode version check.", v => { min_xcode_version = null; }, true /* This is a non-documented option. Please discuss any customers running into the xcode version check on the maciosdev@ list before giving this option out to customers. */);
 			options.Add ("warnaserror:", "An optional comma-separated list of warning codes that should be reported as errors (if no warnings are specified all warnings are reported as errors).", v =>
 			{
 				try {
@@ -386,10 +389,10 @@ namespace Xamarin.Bundler {
 			return System.Reflection.Assembly.GetExecutingAssembly ().Location;
 		}
 
-		static string xcode_bundle_version;
-		public static string XcodeBundleVersion {
+		static string xcode_product_version;
+		public static string XcodeProductVersion {
 			get {
-				return xcode_bundle_version;
+				return xcode_product_version;
 			}
 		}
 
@@ -499,6 +502,100 @@ namespace Xamarin.Bundler {
 			for (int i = 0; i < level; i++)
 				Console.Write ("!");
 			Console.WriteLine ("Timestamp {0}: {1} ms", msg, watch.ElapsedMilliseconds);
+		}
+
+		internal static PDictionary FromPList (string name)
+		{
+			if (!File.Exists (name))
+				throw ErrorHelper.CreateError (24, "Could not find required file '{0}'.", name);
+			return PDictionary.FromFile (name);
+		}
+
+		const string XcodeDefault = "/Applications/Xcode.app";
+
+		static string FindSystemXcode ()
+		{
+			var output = new StringBuilder ();
+			if (Driver.RunCommand ("xcode-select", "-p", output: output) != 0) {
+				ErrorHelper.Warning (59, "Could not find the currently selected Xcode on the system: {0}", output.ToString ());
+				return null;
+			}
+			return output.ToString ().Trim ();
+		}
+
+		static string sdk_root;
+		static string developer_directory;
+
+		public static string DeveloperDirectory {
+			get {
+				return developer_directory;
+			}
+		}
+
+		static void ValidateXcode (bool accept_any_xcode_version, bool warn_if_not_found)
+		{
+			if (sdk_root == null) {
+				sdk_root = FindSystemXcode ();
+				if (sdk_root == null) {
+					// FindSystemXcode showed a warning in this case. In particular do not use 'string.IsNullOrEmpty' here,
+					// because FindSystemXcode may return an empty string (with no warning printed) if the xcode-select command
+					// succeeds, but returns nothing.
+					sdk_root = null;
+				} else if (!Directory.Exists (sdk_root)) {
+					ErrorHelper.Warning (60, "Could not find the currently selected Xcode on the system. 'xcode-select --print-path' returned '{0}', but that directory does not exist.", sdk_root);
+					sdk_root = null;
+				} else {
+					if (!accept_any_xcode_version)
+						ErrorHelper.Warning (61, "No Xcode.app specified (using --sdkroot), using the system Xcode as reported by 'xcode-select --print-path': {0}", sdk_root);
+				}
+				if (sdk_root == null) {
+					sdk_root = XcodeDefault;
+					if (!Directory.Exists (sdk_root)) {
+						if (warn_if_not_found) {
+							// mmp: and now we give up, but don't throw like mtouch, because we don't want to change behavior (this sometimes worked it appears)
+							ErrorHelper.Warning (56, "Cannot find Xcode in any of our default locations. Please install Xcode, or pass a custom path using --sdkroot=<path>.");
+							return; // Can't validate the version below if we can't even find Xcode...
+						}
+
+						throw ErrorHelper.CreateError (56, "Cannot find Xcode in the default location (/Applications/Xcode.app). Please install Xcode, or pass a custom path using --sdkroot <path>.");
+					}
+					ErrorHelper.Warning (62, "No Xcode.app specified (using --sdkroot or 'xcode-select --print-path'), using the default Xcode instead: {0}", sdk_root);
+				}
+			} else if (!Directory.Exists (sdk_root)) {
+				throw ErrorHelper.CreateError (55, "The Xcode path '{0}' does not exist.", sdk_root);
+			}
+
+			// Check what kind of path we got
+			if (File.Exists (Path.Combine (sdk_root, "Contents", "MacOS", "Xcode"))) {
+				// path to the Xcode.app
+				developer_directory = Path.Combine (sdk_root, "Contents", "Developer");
+			} else if (File.Exists (Path.Combine (sdk_root, "..", "MacOS", "Xcode"))) {
+				// path to Contents/Developer
+				developer_directory = Path.GetFullPath (Path.Combine (sdk_root, "..", "..", "Contents", "Developer"));
+			} else {
+				throw ErrorHelper.CreateError (57, "Cannot determine the path to Xcode.app from the sdk root '{0}'. Please specify the full path to the Xcode.app bundle.", sdk_root);
+			}
+			
+			var plist_path = Path.Combine (Path.GetDirectoryName (DeveloperDirectory), "version.plist");
+
+			if (File.Exists (plist_path)) {
+				var plist = FromPList (plist_path);
+				var version = plist.GetString ("CFBundleShortVersionString");
+				xcode_version = new Version (version);
+				xcode_product_version = plist.GetString ("ProductBuildVersion");
+			} else {
+				throw ErrorHelper.CreateError (58, "The Xcode.app '{0}' is invalid (the file '{1}' does not exist).", Path.GetDirectoryName (Path.GetDirectoryName (DeveloperDirectory)), plist_path);
+			}
+
+			if (!accept_any_xcode_version) {
+				if (min_xcode_version != null && XcodeVersion < min_xcode_version)
+					throw ErrorHelper.CreateError (51, "{3} {0} requires Xcode {4} or later. The current Xcode version (found in {2}) is {1}.", Constants.Version, XcodeVersion.ToString (), sdk_root, PRODUCT, min_xcode_version);
+
+				if (XcodeVersion < SdkVersions.XcodeVersion)
+					ErrorHelper.Warning (79, "The recommended Xcode version for {3} {0} is Xcode {3} or later. The current Xcode version (found in {2}) is {1}.", Constants.Version, XcodeVersion.ToString (), sdk_root, SdkVersions.Xcode, PRODUCT);
+			}
+
+			Driver.Log (1, "Using Xcode {0} ({2}) found in {1}", XcodeVersion, sdk_root, XcodeProductVersion);
 		}
 	}
 }
